Refactoring to break span processing out of ot index, synthesizing parent span
