+++ azureLogIngestion/opentelemetry/index.test.ts
@@ -43,63 +43,63 @@ describe("OpenTelemetryAdapter", () => {
 
         expect(adapter).toBeInstanceOf(OpenTelemetryAdapter)
 
-        expect(adapter.spanProcessor).toBeInstanceOf(BatchSpanProcessor)
-        expect(adapter.traceProvider).toBeInstanceOf(NRTracerProvider)
-        expect(adapter.currentBatch).toBeInstanceOf(Array)
-        expect(adapter.getBatchSize()).toEqual(0)
+        expect(adapter.spanProcessor.spanProcessor).toBeInstanceOf(BatchSpanProcessor)
+        expect(adapter.spanProcessor.traceProvider).toBeInstanceOf(NRTracerProvider)
+        expect(adapter.spanProcessor.batch).toBeInstanceOf(Array)
+        expect(adapter.spanProcessor.batch.length).toEqual(0)
     })
 
     it("processes app request and dependency as spans", () => {
         const adapter = new OpenTelemetryAdapter("mock-insert-key", mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(0)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(0)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(appInsightsAppRequest, mockContext)
-        expect(adapter.getBatchSize()).toEqual(1)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(2)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(appInsightsAppDependency, mockContext)
-        expect(adapter.getBatchSize()).toEqual(2)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(2)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
     })
 
     it("processes app request and dependency when sent as array of strings", () => {
         const adapter = new OpenTelemetryAdapter("mock-insert-key", mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(0)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(0)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(arrayOfStrings, mockContext)
-        expect(adapter.getBatchSize()).toEqual(3)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(6)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(appInsightsAppDependency, mockContext)
-        expect(adapter.getBatchSize()).toEqual(4)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(8)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
     })
 
     it("processes app event as span", () => {
         const adapter = new OpenTelemetryAdapter("mock-insert-key", mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(0)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(0)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(appInsightsAppEvent, mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(1)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(2)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
     })
 
     it("processes app exception as span", () => {
         const adapter = new OpenTelemetryAdapter("mock-insert-key", mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(0)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(0)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
 
         adapter.processMessages(appInsightsAppException, mockContext)
 
-        expect(adapter.getBatchSize()).toEqual(1)
-        expect(adapter.currentBatch).toMatchSnapshot()
+        expect(adapter.spanProcessor.batch.length).toEqual(2)
+        expect(adapter.spanProcessor.batch).toMatchSnapshot()
     })
 })
