+++ azureLogIngestion/opentelemetry/processors/index.ts
@@ -1,0 +1,1 @@
+export { default as SpanProcessor } from "./spans"
